Move format_ange to util namespace and file structure
