+++ standalone/include/psen_scan_v2_standalone/protocol_layer/scanner_state_machine.h
@@ -36,7 +36,7 @@
 #include "psen_scan_v2_standalone/protocol_layer/scanner_events.h"
 
 #include "psen_scan_v2_standalone/logging.h"
-#include "psen_scan_v2_standalone/format_range.h"
+#include "psen_scan_v2_standalone/util/format_range.h"
 #include "psen_scan_v2_standalone/communication_layer/udp_client.h"
 
 #include "psen_scan_v2_standalone/api/laserscan.h"
